ui/requirements.txt
streamlit==1.39.0
PyYAML==6.0.1
Pillow==8.3.2